ex01 done - cout<< overload and better text formatting

+++ 02/ex01/src/main.cpp
@@ -8,11 +8,11 @@ int	main(void)
 	Fixed const c(42.42f);
 	Fixed const d(b);
 
-	a = Fixed( 1234.4321f );
-	std::cout << "a is " << a.toFloat() << std::endl;
-	std::cout << "b is " << b.toFloat() << std::endl;
-	std::cout << "c is " << c.toFloat() << std::endl;
-	std::cout << "d is " << d.toFloat() << std::endl;
+	a = Fixed(1234.4321f);
+	std::cout << "a is " << a << std::endl;
+	std::cout << "b is " << b << std::endl;
+	std::cout << "c is " << c << std::endl;
+	std::cout << "d is " << d << std::endl;
 	std::cout << "a is " << a.toInt() << " as integer" << std::endl;
 	std::cout << "b is " << b.toInt() << " as integer" << std::endl;
 	std::cout << "c is " << c.toInt() << " as integer" << std::endl;
